Fixed int/double mismatch.

+++ app/src/main/java/com/example/stacy/refill/Calculation/MACalculation.java
@@ -2,10 +2,10 @@ package com.example.stacy.refill.Calculation;
 
 public class MACalculation {
 
-    public static int period = 7;
+    public static double period = 7d;
 
     public static double Calculate(double prevMA, double newDays){
-        double alpha = 2 / (1 + period);
+        double alpha = 2d / (1d + period);
         return newDays * alpha + prevMA * (1 - alpha);
     }
 }
